+++ changelog.md
@@ -1,6 +1,7 @@
 v2.1.2
 ------
 * Add Zephyr Engine, to generate wind tunnels
+* Fix issue with wand transform mod compatibility
 
 v2.1.1
 ------